refactor: improve typing of decorators and remove redundant module
