+++ sourcery_analytics/metrics/cyclomatic_complexity.py
@@ -9,8 +9,7 @@ For implementation details see methods below.
 import astroid
 
 from sourcery_analytics.conditions import is_elif
-from sourcery_analytics.utils import nodedispatch
-from sourcery_analytics.validators import validate_node_type
+from sourcery_analytics.utils import nodedispatch, validate_node_type
 from sourcery_analytics.visitors import TreeVisitor, FunctionVisitor
