+++ sourcery_analytics/metrics/method_length.py
@@ -7,8 +7,7 @@ Note that for obvious reasons, the method length is only defined for methods.
 """
 import astroid.nodes
 
-from sourcery_analytics.utils import nodedispatch
-from sourcery_analytics.validators import validate_node_type
+from sourcery_analytics.utils import nodedispatch, validate_node_type
 from sourcery_analytics.visitors import TreeVisitor, FunctionVisitor
